py/glm-ocr/other/pyproject.toml
[project]
name = "glm-ocr-other"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "glm-ocr",
    "onnxruntime>=1.24.2",
    "onnx",
    "pillow",
    "numpy",
    "ml-dtypes",
]

[project.optional-dependencies]
export = [
    "torch>=2.6,<2.10",
    "torchvision",
    "huggingface-hub",
]

[tool.uv.sources]
glm-ocr = { path = "..", editable = true }
